Club-Hub project/club-hub-backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
drf-yasg>=1.21.7
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0 